+++ relay-event-normalization/src/eap/mod.rs
@@ -361,7 +361,15 @@ pub fn normalize_user_geo(
 }
 
 /// Normalizes the [DSC](DynamicSamplingContext) into [`Attributes`].
-pub fn normalize_dsc(attributes: &mut Annotated<Attributes>, dsc: Option<&DynamicSamplingContext>) {
+///
+/// If `is_segment` is set to `false`, the function will only add select attributes that are
+/// necessary on every span - both segment and non-segment - for dynamic sampling to work. More
+/// attributes are added when `is_segment` is set to `true`.
+pub fn normalize_dsc(
+    attributes: &mut Annotated<Attributes>,
+    dsc: Option<&DynamicSamplingContext>,
+    is_segment: bool,
+) {
     let Some(dsc) = dsc else { return };
 
     let attributes = attributes.get_or_insert_with(Default::default);
@@ -370,23 +378,26 @@ pub fn normalize_dsc(attributes: &mut Annotated<Attributes>, dsc: Option<&Dynami
     if attributes.contains_key(SENTRY__DSC__TRACE_ID) {
         return;
     }
-
     attributes.insert(SENTRY__DSC__TRACE_ID, dsc.trace_id.to_string());
-    attributes.insert(SENTRY__DSC__PUBLIC_KEY, dsc.public_key.to_string());
-    if let Some(release) = &dsc.release {
-        attributes.insert(SENTRY__DSC__RELEASE, release.clone());
-    }
-    if let Some(environment) = &dsc.environment {
-        attributes.insert(SENTRY__DSC__ENVIRONMENT, environment.clone());
-    }
+
     if let Some(transaction) = &dsc.transaction {
         attributes.insert(SENTRY__DSC__TRANSACTION, transaction.clone());
     }
-    if let Some(sample_rate) = dsc.sample_rate {
-        attributes.insert(SENTRY__DSC__SAMPLE_RATE, sample_rate);
-    }
-    if let Some(sampled) = dsc.sampled {
-        attributes.insert(SENTRY__DSC__SAMPLED, sampled);
+
+    if is_segment {
+        attributes.insert(SENTRY__DSC__PUBLIC_KEY, dsc.public_key.to_string());
+        if let Some(release) = &dsc.release {
+            attributes.insert(SENTRY__DSC__RELEASE, release.clone());
+        }
+        if let Some(environment) = &dsc.environment {
+            attributes.insert(SENTRY__DSC__ENVIRONMENT, environment.clone());
+        }
+        if let Some(sample_rate) = dsc.sample_rate {
+            attributes.insert(SENTRY__DSC__SAMPLE_RATE, sample_rate);
+        }
+        if let Some(sampled) = dsc.sampled {
+            attributes.insert(SENTRY__DSC__SAMPLED, sampled);
+        }
     }
 }
 
@@ -726,9 +737,88 @@ pub fn write_legacy_attributes(attributes: &mut Annotated<Attributes>) {
 
 #[cfg(test)]
 mod tests {
-    use relay_protocol::{Empty, SerializableAnnotated};
+    use relay_protocol::{Empty, SerializableAnnotated, assert_annotated_snapshot};
 
     use super::*;
+
+    fn mock_dsc(transaction: Option<&str>) -> DynamicSamplingContext {
+        DynamicSamplingContext {
+            trace_id: "67e5504410b1426f9247bb680e5fe0c8".parse().unwrap(),
+            public_key: "12345678901234567890123456789012".parse().unwrap(),
+            release: None,
+            environment: None,
+            transaction: transaction.map(str::to_owned),
+            sample_rate: None,
+            user: Default::default(),
+            replay_id: None,
+            sampled: None,
+            other: Default::default(),
+        }
+    }
+
+    #[test]
+    fn test_normalize_dsc_child_span_no_dsc() {
+        let mut attributes = Annotated::empty();
+        normalize_dsc(&mut attributes, None, false);
+        assert!(attributes.value().is_none());
+    }
+
+    #[test]
+    fn test_normalize_dsc_child_span_no_transaction() {
+        let mut attributes = Annotated::empty();
+        let dsc = mock_dsc(None);
+        normalize_dsc(&mut attributes, Some(&dsc), false);
+        assert_annotated_snapshot!(attributes, @r#"
+        {
+          "sentry.dsc.trace_id": {
+            "type": "string",
+            "value": "67e5504410b1426f9247bb680e5fe0c8"
+          }
+        }
+        "#);
+    }
+
+    #[test]
+    fn test_normalize_dsc_child_span() {
+        let mut attributes = Annotated::empty();
+        let dsc = mock_dsc(Some("/some/endpoint"));
+        normalize_dsc(&mut attributes, Some(&dsc), false);
+        assert_annotated_snapshot!(attributes, @r#"
+        {
+          "sentry.dsc.trace_id": {
+            "type": "string",
+            "value": "67e5504410b1426f9247bb680e5fe0c8"
+          },
+          "sentry.dsc.transaction": {
+            "type": "string",
+            "value": "/some/endpoint"
+          }
+        }
+        "#);
+    }
+
+    #[test]
+    fn test_normalize_dsc_segment() {
+        let mut attributes = Annotated::empty();
+        let dsc = mock_dsc(Some("/some/endpoint"));
+        normalize_dsc(&mut attributes, Some(&dsc), true);
+        assert_annotated_snapshot!(attributes, @r#"
+        {
+          "sentry.dsc.public_key": {
+            "type": "string",
+            "value": "12345678901234567890123456789012"
+          },
+          "sentry.dsc.trace_id": {
+            "type": "string",
+            "value": "67e5504410b1426f9247bb680e5fe0c8"
+          },
+          "sentry.dsc.transaction": {
+            "type": "string",
+            "value": "/some/endpoint"
+          }
+        }
+        "#);
+    }
 
     #[test]
     fn test_normalize_received_none() {
@@ -739,7 +829,7 @@ mod tests {
             DateTime::from_timestamp_nanos(1_234_201_337),
         );
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "sentry.observed_timestamp_nanos": {
             "type": "string",
@@ -766,7 +856,7 @@ mod tests {
             DateTime::from_timestamp_nanos(1_234_201_337),
         );
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r###"
+        assert_annotated_snapshot!(attributes, @r###"
         {
           "sentry.observed_timestamp_nanos": {
             "type": "string",
@@ -863,7 +953,7 @@ mod tests {
         let mut attributes = Annotated::<Attributes>::from_json(json).unwrap();
         normalize_attribute_types(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "double_with_i64": {
             "type": "double",
@@ -1035,7 +1125,7 @@ mod tests {
             }),
         );
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "browser.name": {
             "type": "string",
@@ -1075,7 +1165,7 @@ mod tests {
             }),
         );
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "browser.name": {
             "type": "string",
@@ -1086,7 +1176,7 @@ mod tests {
             "value": "13.3.7"
           }
         }
-        "#,
+        "#
         );
     }
 
@@ -1112,7 +1202,7 @@ mod tests {
             })
         });
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "client.address": {
             "type": "string",
@@ -1152,7 +1242,7 @@ mod tests {
 
         normalize_user_geo(&mut attributes, |_| unreachable!());
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "client.address": {
             "type": "string",
@@ -1163,7 +1253,7 @@ mod tests {
             "value": "Foo Hausen"
           }
         }
-        "#,
+        "#
         );
     }
 
@@ -1226,7 +1316,7 @@ mod tests {
 
         normalize_attribute_names_inner(&mut attributes, mock_attribute_info);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r###"
+        assert_annotated_snapshot!(attributes, @r###"
         {
           "backfill.empty": {
             "type": "string",
@@ -1299,7 +1389,7 @@ mod tests {
 
         normalize_sentry_op(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "db.operation.name": {
             "type": "string",
@@ -1345,7 +1435,7 @@ mod tests {
 
         normalize_db_attributes(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "db.operation.name": {
             "type": "string",
@@ -1411,7 +1501,7 @@ mod tests {
 
         normalize_db_attributes(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "db.collection.name": {
             "type": "string",
@@ -1537,7 +1627,7 @@ mod tests {
 
         normalize_db_attributes(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "http.request.method": {
             "type": "string",
@@ -1583,7 +1673,7 @@ mod tests {
 
         normalize_http_attributes(&mut attributes, &[]);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "http.request.method": {
             "type": "string",
@@ -1641,7 +1731,7 @@ mod tests {
 
         normalize_http_attributes(&mut attributes, &[]);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "http.request.method": {
             "type": "string",
@@ -1702,7 +1792,7 @@ mod tests {
             &["application.www.xn--85x722f.xn--55qx5d.cn".to_owned()],
         );
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "http.request.method": {
             "type": "string",
@@ -1760,7 +1850,7 @@ mod tests {
 
         normalize_db_attributes(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "db.collection.name": {
             "type": "string",
@@ -1819,7 +1909,7 @@ mod tests {
         normalize_attribute_names(&mut attributes);
         normalize_http_attributes(&mut attributes, &[]);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "http.request.method": {
             "type": "string",
@@ -1865,7 +1955,7 @@ mod tests {
 
         normalize_http_attributes(&mut attributes, &[]);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "http.request.method": {
             "type": "string",
@@ -1935,7 +2025,7 @@ mod tests {
 
         write_legacy_attributes(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r###"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "db.collection.name": {
             "type": "string",
@@ -1982,7 +2072,7 @@ mod tests {
             "value": "db"
           }
         }
-        "###);
+        "#);
     }
 
     #[test]
@@ -2004,7 +2094,7 @@ mod tests {
 
         normalize_span_category(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "sentry.category": {
             "type": "string",
@@ -2032,7 +2122,7 @@ mod tests {
 
         normalize_span_category(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "sentry.category": {
             "type": "string",
@@ -2060,7 +2150,7 @@ mod tests {
 
         normalize_span_category(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "sentry.category": {
             "type": "string",
@@ -2088,7 +2178,7 @@ mod tests {
 
         normalize_span_category(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "sentry.category": {
             "type": "string",
@@ -2117,7 +2207,7 @@ mod tests {
 
         normalize_span_category(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "db.system.name": {
             "type": "string",
@@ -2146,7 +2236,7 @@ mod tests {
 
         normalize_span_category(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "http.request.method": {
             "type": "string",
@@ -2175,7 +2265,7 @@ mod tests {
 
         normalize_span_category(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "sentry.category": {
             "type": "string",
@@ -2204,7 +2294,7 @@ mod tests {
 
         normalize_span_category(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "resource.render_blocking_status": {
             "type": "string",
@@ -2233,7 +2323,7 @@ mod tests {
 
         normalize_span_category(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "sentry.category": {
             "type": "string",
@@ -2261,7 +2351,7 @@ mod tests {
 
         normalize_client_address(&mut attributes, Some("192.168.1.1".parse().unwrap()));
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "client.address": {
             "type": "string",
@@ -2285,7 +2375,7 @@ mod tests {
 
         normalize_client_address(&mut attributes, None);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {}
         "#);
     }
@@ -2304,7 +2394,7 @@ mod tests {
 
         normalize_client_address(&mut attributes, Some("192.168.1.1".parse().unwrap()));
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "client.address": {
             "type": "string",
@@ -2337,7 +2427,7 @@ mod tests {
 
         normalize_client_address(&mut attributes, Some("2001:db8::1".parse().unwrap()));
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "client.address": {
             "type": "string",
@@ -2353,7 +2443,7 @@ mod tests {
 
         normalize_inject_client_address(&mut attributes, Some("192.168.1.1".parse().unwrap()));
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "client.address": {
             "type": "string",
@@ -2377,7 +2467,7 @@ mod tests {
 
         normalize_inject_client_address(&mut attributes, Some("192.168.1.1".parse().unwrap()));
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "client.address": {
             "type": "string",
@@ -2393,7 +2483,7 @@ mod tests {
 
         normalize_inject_client_address(&mut attributes, None);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {}
         "#);
     }
@@ -2404,7 +2494,7 @@ mod tests {
 
         normalize_inject_client_address(&mut attributes, Some("2001:db8::1".parse().unwrap()));
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "client.address": {
             "type": "string",
@@ -2429,7 +2519,7 @@ mod tests {
 
         normalize_span_category(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "some.other.attribute": {
             "type": "string",